+++ CHANGELOG.md
@@ -16,6 +16,10 @@
 
 - experimental support of tagged template literals, eg ```sql`select * from ${table}```. https://github.com/rescript-lang/rescript-compiler/pull/6250
 
+#### :bug: Bug Fix
+
+- GenType: distinguish inline records from unary variant cases of object type. https://github.com/rescript-lang/rescript-compiler/pull/6586
+
 # 11.0.1
 
 #### :bug: Bug Fix